Add Microtask Fix Opt-Out "option" for CEP platforms
Certain versions of CEP 11 have the same issue as 9.3 and earlier. Additionally, there are some situations where a CEP user may want to explicitly opt-out of the Microtask Fix logic. This commit enables this change by allowing a developer to define a `__vue_microtask_fix_optout__` property on the `__adobe_cep__` global object that is present on CEP environments. If that property exists when Vue.js is imported, then the Microtask Fix logic will be skipped entirely.